+++ src/test/groovy/org/jenkinsci/gradle/plugins/jpi/JpiPluginSpec.groovy
@@ -1,6 +1,8 @@
 package org.jenkinsci.gradle.plugins.jpi
 
 import org.gradle.api.Project
+import org.gradle.api.internal.project.ProjectInternal
+import org.gradle.api.plugins.JavaPluginConvention
 import org.gradle.testfixtures.ProjectBuilder
 import spock.lang.Specification
 
@@ -26,11 +28,37 @@ class JpiPluginSpec extends Specification {
         taskClass.isInstance(project.tasks[taskName])
 
         where:
-        taskName    |  taskClass
+        taskName     | taskClass
         'jpi'        | Jpi
         'server'     | ServerTask
         'localizer'  | LocalizerTask
         'stapler'    | StaplerGroovyStubsTask
         'insertTest' | TestInsertionTask
+    }
+
+    def 'localizer task has been setup'(Object outputDir, String expectedOutputDir) {
+        when:
+        project.with {
+            apply plugin: 'jpi'
+            jenkinsPlugin {
+                localizerOutputDir = outputDir
+            }
+        }
+        (project as ProjectInternal).projectEvaluationBroadcaster.afterEvaluate(project, project.state)
+
+        then:
+        JavaPluginConvention javaConvention = project.convention.getPlugin(JavaPluginConvention)
+        LocalizerTask localizerTask = project.tasks.findByName('localizer') as LocalizerTask
+        localizerTask != null
+        localizerTask.destinationDir == new File(project.rootDir, expectedOutputDir)
+        localizerTask.sourceDirs == javaConvention.sourceSets.main.resources.srcDirs
+        project.tasks.compileJava.dependsOn.contains(localizerTask)
+        javaConvention.sourceSets.main.java.srcDirs.contains(localizerTask.destinationDir)
+
+        where:
+        outputDir       | expectedOutputDir
+        null            | 'build/generated-src/localizer'
+        'foo'           | 'foo'
+        new File('bar') | 'bar'
     }
 }